+++ src/main/java/com/iexec/blockchain/command/task/contribute/TaskContributeRepository.java
@@ -1,6 +1,0 @@
-package com.iexec.blockchain.command.task.contribute;
-
-import com.iexec.blockchain.command.generic.CommandRepository;
-
-public interface TaskContributeRepository extends CommandRepository<TaskContribute> {
-}